use channels for fixed permits

+++ src/dbnode/server/server.go
@@ -474,7 +474,7 @@ func Run(runOpts RunOptions) {
 
 	permitOpts := opts.PermitsOptions().SetSeriesReadPermitsManager(seriesReadPermits)
 	if cfg.Index.MaxQueryIDsConcurrency != 0 {
-		permitOpts.SetIndexQueryPermitsManager(permits.NewFixedPermitsManager(int64(cfg.Index.MaxQueryIDsConcurrency)))
+		permitOpts.SetIndexQueryPermitsManager(permits.NewFixedPermitsManager(cfg.Index.MaxQueryIDsConcurrency))
 	} else {
 		logger.Warn("max index query IDs concurrency was not set, falling back to default value")
 	}
